+++ src/components/CheckForm/MultiHttpNew.test.tsx
@@ -7,7 +7,7 @@ import { server } from 'test/server';
 
 import { AlertSensitivity, CheckType, ROUTES } from 'types';
 import { selectOption, submitForm } from 'components/CheckEditor/testHelpers';
-import { PLUGIN_URL_PATH } from 'components/constants';
+import { FIVE_MINUTES_IN_MS, PLUGIN_URL_PATH } from 'components/constants';
 
 import { CheckForm } from './CheckForm';
 
@@ -62,6 +62,9 @@ describe('new checks', () => {
     // Set probe options
     await selectOption(user, { label: 'Probe locations', option: PUBLIC_PROBE.name });
 
+    // Open requests section
+    await user.click(screen.getByRole('button', { name: /Requests/ }));
+
     // Add a custom label
     const addCustomLabelButton = await screen.findByRole('button', { name: /Add label/ });
     await user.click(addCustomLabelButton);
@@ -106,7 +109,7 @@ describe('new checks', () => {
       alertSensitivity: AlertSensitivity.None,
       basicMetricsOnly: true,
       enabled: true,
-      frequency: 120000,
+      frequency: FIVE_MINUTES_IN_MS,
       job: JOB,
       labels: LABELS,
       probes: [PUBLIC_PROBE.id],