+++ src/components/Gauges/SuccessRateGaugeCheckUptime.tsx
@@ -14,7 +14,8 @@ type SuccessRateGaugeCheckUptimeProps = {
 
 export const SuccessRateGaugeCheckUptime = ({ check, height, width, onClick }: SuccessRateGaugeCheckUptimeProps) => {
   const { data, isLoading, isFetching } = useCheckUptimeSuccessRate(check);
-  const value = data ? data.value[1] : null;
+
+  const value = data ? data[0]?.value?.[1] : null;
 
   return (
     <Gauge