+++ src/components/SliderInput.tsx
@@ -1,19 +1,18 @@
-import React, { useEffect } from 'react';
-import { useFormContext } from 'react-hook-form';
+import React from 'react';
+import { Controller, useFormContext } from 'react-hook-form';
 import { GrafanaTheme2 } from '@grafana/data';
-import { Slider, useStyles2 } from '@grafana/ui';
+import { useStyles2 } from '@grafana/ui';
 import { css } from '@emotion/css';
 
-interface Props {
-  defaultValue: number;
+import { TimeSlider } from './TimeSlider/TimeSlider';
+
+export interface SliderInputProps {
   min: number;
   max: number;
   name: string;
   id?: string;
   validate?: (value: number) => string | undefined;
-  prefixLabel?: string;
   step?: number;
-  suffixLabel?: string;
   invalid?: boolean;
 }
 
@@ -23,7 +22,8 @@ const getStyles = (theme: GrafanaTheme2) => ({
     align-items: center;
   `,
   slider: css`
-    width: 250px;
+    width: 100%;
+    max-width: 750px;
     margin-right: ${theme.spacing(2)};
     margin-left: ${theme.spacing(1)};
   `,
@@ -35,43 +35,25 @@ const getStyles = (theme: GrafanaTheme2) => ({
   rightMargin: css`
     margin-right: 0.5rem;
   `,
-  sliderInput: css`
-    width: 40px;
-    margin-right: 0.5rem;
-  `,
 });
 
-export const SliderInput = ({ min, max, prefixLabel, suffixLabel, name, step = 1, validate, defaultValue }: Props) => {
+export const SliderInput = ({ min, max, name, step = 1, validate }: SliderInputProps) => {
   const styles = useStyles2(getStyles);
-  const { register, setValue, setError, getValues, clearErrors } = useFormContext(); // TODO: type correctly
-
-  useEffect(() => {
-    register(name);
-    setValue(name, defaultValue);
-  }, [name, register, defaultValue, setValue]);
+  const { control } = useFormContext();
 
   return (
     <div className={styles.container} data-testid={name}>
-      {prefixLabel}
       <div className={styles.slider}>
-        <Slider
-          tooltipAlwaysVisible={false}
-          min={min ?? 0}
-          max={max}
-          step={step}
-          value={getValues(name)}
-          onChange={(value) => {
-            const error = validate && validate(value);
-            if (error) {
-              setError(name, { message: error });
-            } else {
-              clearErrors(name);
-              setValue(name, value);
-            }
+        <Controller
+          name={name}
+          control={control}
+          rules={{ validate: (v) => validate?.(v) }}
+          render={({ field }) => {
+            const { ref, ...rest } = field;
+            return <TimeSlider {...rest} min={min ?? 0} max={max} step={step} />;
           }}
         />
       </div>
-      {suffixLabel}
     </div>
   );
 };