+++ src/scenes/TCP/getTcpScene.ts
@@ -27,6 +27,7 @@ import { getAlertAnnotations } from 'scenes/Common/alertAnnotations';
 import { getEditButton } from 'scenes/Common/editButton';
 import { getEmptyScene } from 'scenes/Common/emptyScene';
 import { getErrorRateTimeseries } from 'scenes/HTTP/errorRateTimeseries';
+import { getMinStepFromFrequency } from 'scenes/utils';
 
 export function getTcpScene({ metrics, logs, singleCheckMode }: DashboardSceneAppConfig, checks: Check[]) {
   return () => {
@@ -42,11 +43,12 @@ export function getTcpScene({ metrics, logs, singleCheckMode }: DashboardSceneAp
     const { job, instance, probe } = getVariables(CheckType.TCP, metrics, checks, singleCheckMode);
 
     const variables = new SceneVariableSet({ variables: [probe, job, instance] });
-    const errorMap = getErrorRateMapPanel(metrics);
 
-    const uptime = getUptimeStat(metrics);
-    const reachability = getReachabilityStat(metrics);
-    const avgLatency = getAvgLatencyStat(metrics);
+    const minStep = getMinStepFromFrequency(checks?.[0]?.frequency);
+    const errorMap = getErrorRateMapPanel(metrics, minStep);
+    const uptime = getUptimeStat(metrics, minStep);
+    const reachability = getReachabilityStat(metrics, minStep);
+    const avgLatency = getAvgLatencyStat(metrics, minStep);
     const sslExpiry = getSSLExpiryStat(metrics);
     const frequency = getFrequencyStat(metrics);
 
@@ -57,7 +59,7 @@ export function getTcpScene({ metrics, logs, singleCheckMode }: DashboardSceneAp
       }),
     });
 
-    const errorRateTimeseries = getErrorRateTimeseries(metrics);
+    const errorRateTimeseries = getErrorRateTimeseries(metrics, minStep);
     const topRight = new SceneFlexLayout({
       direction: 'column',
       children: [new SceneFlexItem({ height: 90, body: statRow }), new SceneFlexItem({ body: errorRateTimeseries })],